+++ eslint-config/test/data/good.vue
@@ -1,0 +1,15 @@
+<template>
+    <div class="nice">
+        <p>Here is a message from our sponsor: {{ msg }}</p>
+    </div>
+</template>
+
+<script>
+const component = {
+    data: () => ({
+        msg: 'Buy our stuff!',
+    }),
+};
+
+export default component;
+</script>